COMMON-3801: Refine the error code
