+++ common/exception/tests/it/prelude.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common_exception::prelude::*;
+use common_exception::*;
 
 #[test]
 fn test_prelude() -> anyhow::Result<()> {
