+++ metasrv/src/meta_service/raftmeta.rs
@@ -25,8 +25,8 @@ use common_base::tokio::sync::watch;
 use common_base::tokio::sync::Mutex;
 use common_base::tokio::sync::RwLockReadGuard;
 use common_base::tokio::task::JoinHandle;
-use common_exception::prelude::ErrorCode;
-use common_exception::prelude::ToErrorCode;
+use common_exception::ErrorCode;
+use common_exception::ToErrorCode;
 use common_meta_api::MetaApi;
 use common_meta_raft_store::config::RaftConfig;
 use common_meta_raft_store::state_machine::StateMachine;
