+++ query/tests/it/api/rpc/flight_service.rs
@@ -20,9 +20,9 @@ use common_arrow::arrow_format::flight::data::Ticket;
 use common_arrow::arrow_format::flight::service::flight_service_server::FlightService;
 use common_base::tokio;
 use common_datavalues::DataValue;
-use common_exception::exception::ABORT_SESSION;
 use common_exception::ErrorCode;
 use common_exception::Result;
+use common_exception::ABORT_SESSION;
 use common_planners::Expression;
 use databend_query::api::DatabendQueryFlightDispatcher;
 use databend_query::api::DatabendQueryFlightService;
